Use derive color function to get theme color dark
